pull out garboStrategy into combatStrategy.ts
