+++ packages/garbo/src/tasks/freeFight.ts
@@ -70,7 +70,8 @@ import {
   Witchess,
 } from "libram";
 import { acquire } from "../acquire";
-import { GarboStrategy, Macro } from "../combat";
+import { Macro } from "../combat";
+import { GarboStrategy } from "../combatStrategy";
 import { globalOptions } from "../config";
 import { garboValue } from "../garboValue";
 import { freeFightOutfit } from "../outfit";
