+++ packages/garbo/src/tasks/freeGiantSandworm.ts
@@ -31,7 +31,8 @@ import {
   sum,
   uneffect,
 } from "libram";
-import { GarboStrategy, Macro } from "../combat";
+import { Macro } from "../combat";
+import { GarboStrategy } from "../combatStrategy";
 import { globalOptions } from "../config";
 import { garboValue } from "../garboValue";
 import { freeFightOutfit } from "../outfit";
