+++ packages/garbo/src/tasks/post/index.ts
@@ -40,7 +40,8 @@ import {
   uneffect,
   withProperty,
 } from "libram";
-import { GarboStrategy, Macro } from "../../combat";
+import { Macro } from "../../combat";
+import { GarboStrategy } from "../../combatStrategy";
 import { globalOptions } from "../../config";
 import { computeDiet, consumeDiet } from "../../diet";
 import {
